Add typemap for checking the values of the enum.
The current ability to throw any int into a method that is expecting an enumeration is a problem for C++.  It seems the standard currently defines casting to a value outside the value of an enumeration as resulting in a value that is unspecified.  Not very helpful, but in the future this will be change and will result in undefined behaviour.  So letting this go unchecked is not going to be an option.